Developer

+++ usuarios/bd_create/clientes-guardar.php
@@ -6,7 +6,7 @@ $consultaZona=mysqli_query($conexionBdAdmin,"SELECT * FROM localidad_ciudades WH
 $zona = mysqli_fetch_array($consultaZona, MYSQLI_BOTH);
 
 if (trim($_POST["usuario"]) != "") {
-    $consultaClientesV=mysqli_query($conexionBdPrincipal,"SELECT * FROM clientes WHERE cli_usuario='" . trim($_POST["usuario"]) . "'");
+    $consultaClientesV=mysqli_query($conexionBdPrincipal,"SELECT * FROM clientes WHERE cli_usuario='" . trim($_POST["usuario"]) . "' AND cli_id_empresa='".$idEmpresa."'");
     $clienteV = mysqli_num_rows($consultaClientesV);
     if ($clienteV > 0) {
         echo "<div style='font-family:arial; text-align:center'>Ya existe un cliente con este n&uacute;mero de NIT. Verifique para que no lo registre nuevamente.<br><br>
@@ -31,14 +31,17 @@ $clave1 = generarClaves();
 $clave2 = generarClaves();
 
 $direccion = $_POST["op1"] . " " . $_POST["op2"] . " " . $_POST["op3"] . " # " . $_POST["op4"] . " " . $_POST["op5"] . " - " . $_POST["op6"] . " - " . $_POST["op7"];
-mysqli_query($conexionBdPrincipal,"INSERT INTO clientes(cli_nombre, cli_referencia, cli_categoria, cli_email, cli_telefono, cli_ciudad, cli_usuario, cli_clave, cli_direccion, cli_zona, cli_fecha_registro, cli_fecha_ingreso, cli_nivel, cli_celular, cli_telefonos, cli_sigla, cli_responsable, cli_clave_documentos, cli_tipo_documento, cli_pais, cli_ciudad_extranjera)VALUES('" . $_POST["nombre"] . "','" . $_POST["referencia"] . "','" . $_POST["categoria"] . "','" . $_POST["email"] . "','" . $_POST["telefono"] . "','" . $_POST["ciudad"] . "','" . trim($_POST["usuario"]) . "','" . $clave1 . "','" . strtoupper($direccion) . "','" . $zona[2] . "',now(),'" . $_POST["fechaIngreso"] . "','" . $_POST["nivel"] . "','" . $_POST["celular"] . "','" . $_POST["telefonos"] . "','" . $_POST["sigla"] . "','" . $_SESSION["id"] . "','" . $clave2 . "','" . $_POST["tipoDocumento"] . "','" . $pais . "','" . $city . "')");
+mysqli_query($conexionBdPrincipal,"INSERT INTO clientes(cli_nombre, cli_referencia, cli_categoria, cli_email, cli_telefono, cli_ciudad, cli_usuario, cli_clave, cli_direccion, cli_zona, cli_fecha_registro, cli_fecha_ingreso, cli_nivel, cli_celular, cli_telefonos, cli_sigla, cli_responsable, cli_clave_documentos, cli_tipo_documento, cli_pais, cli_ciudad_extranjera, cli_id_empresa, cli_usuario_acceso)VALUES('" . $_POST["nombre"] . "','" . $_POST["referencia"] . "','" . $_POST["categoria"] . "','" . $_POST["email"] . "','" . $_POST["telefono"] . "','" . $_POST["ciudad"] . "','" . trim($_POST["usuario"]) . "','" . $clave1 . "','" . strtoupper($direccion) . "','" . $zona[2] . "',now(),'" . $_POST["fechaIngreso"] . "','" . $_POST["nivel"] . "','" . $_POST["celular"] . "','" . $_POST["telefonos"] . "','" . $_POST["sigla"] . "','" . $_SESSION["id"] . "','" . $clave2 . "','" . $_POST["tipoDocumento"] . "','" . $pais . "','" . $city . "','" . $idEmpresa . "','" . trim($_POST["usuarioAcceso"]) . "".$_POST["dominio"]."')");
 $idInsertU = mysqli_insert_id($conexionBdPrincipal);
-$numero = (count($_POST["grupos"]));
-$contador = 0;
-mysqli_query($conexionBdPrincipal,"DELETE FROM clientes_categorias WHERE cpcat_cliente='" . $idInsertU . "'");
-while ($contador < $numero) {
-    mysqli_query($conexionBdPrincipal,"INSERT INTO clientes_categorias(cpcat_cliente, cpcat_categoria)VALUES('" . $idInsertU . "'," . $_POST["grupos"][$contador] . ")");
-    $contador++;
+
+if( !empty($_POST["grupos"]) ){
+    $numero = (count($_POST["grupos"]));
+    $contador = 0;
+    mysqli_query($conexionBdPrincipal,"DELETE FROM clientes_categorias WHERE cpcat_cliente='" . $idInsertU . "'");
+    while ($contador < $numero) {
+        mysqli_query($conexionBdPrincipal,"INSERT INTO clientes_categorias(cpcat_cliente, cpcat_categoria)VALUES('" . $idInsertU . "'," . $_POST["grupos"][$contador] . ")");
+        $contador++;
+    }
 }
 
 //Crear automáticamente la sucursal
