+++ usuarios/facturacion.php
@@ -68,7 +68,9 @@ include("includes/head.php");
             <?php include("includes/notificaciones.php");?>
             <p>
             <a href="javascript:history.go(-1);" class="btn btn-primary"><i class="icon-arrow-left"></i> Regresar</a>
-            <a href="facturacion-agregar.php?cte=<?=$_GET["cte"];?>" class="btn btn-danger"><i class="icon-plus"></i> Agregar nuevo</a>
+			<?php if( Modulos::validarRol(['260'], $conexionBdPrincipal, $conexionBdAdmin, $datosUsuarioActual, $configuracion) ) {?>
+            	<a href="facturacion-agregar.php?cte=<?=$_GET["cte"];?>" class="btn btn-danger"><i class="icon-plus"></i> Agregar nuevo</a>
+			<?php }?>
             </p>
 			<div class="row-fluid">
 				<div class="span12">
@@ -94,12 +96,13 @@ include("includes/head.php");
 								INNER JOIN clientes ON cli_id=fact_cliente
 								INNER JOIN ".BDADMIN.".localidad_ciudades ON ciu_id=cli_ciudad
 								INNER JOIN ".BDADMIN.".localidad_departamentos ON dep_id=ciu_departamento
-								WHERE fact_cliente='".$_GET["cte"]."'");
+								WHERE fact_cliente='".$_GET["cte"]."' AND fact_id_empresa='".$idEmpresa."'");
 							}else{
 								$consulta = mysqli_query($conexionBdPrincipal,"SELECT * FROM facturacion
 								INNER JOIN clientes ON cli_id=fact_cliente
 								INNER JOIN ".BDADMIN.".localidad_ciudades ON ciu_id=cli_ciudad
 								INNER JOIN ".BDADMIN.".localidad_departamentos ON dep_id=ciu_departamento
+								WHERE  fact_id_empresa='".$idEmpresa."'
 								");
 							}
 							$no = 1;
@@ -120,7 +123,7 @@ include("includes/head.php");
 								
 								$saldoFinal = $valorReal - $abonos[0];
 
-								if($datosUsuarioActual[3]!=1){
+								if(!Modulos::validarRol([383], $conexionBdPrincipal, $conexionBdAdmin, $datosUsuarioActual, $configuracion)){
 									$consultaNumZonas=mysqli_query($conexionBdPrincipal,"SELECT * FROM zonas_usuarios WHERE zpu_usuario='".$_SESSION["id"]."' AND zpu_zona='".$res['cli_zona']."'");
 									$numZ = mysqli_num_rows($consultaNumZonas);
 									if($numZ==0) continue;
@@ -135,15 +138,23 @@ include("includes/head.php");
 									<?php echo "<b>Nombre</b>: ". $res['cli_nombre']."<br>";?>
                                     <?php echo "<b>Ciudad</b>: ". $res['ciu_nombre'].", ".$res['dep_nombre']." (<b>03".$res['dep_indicativo']."</b>)";?>
                                     <h4 style="margin-top:10px;">
-                                	<a href="facturacion-editar.php?id=<?=$res['fact_id'];?>&cte=<?=$_GET['cte'];?>" data-toggle="tooltip" title="Editar"><i class="icon-edit"></i></a>&nbsp;
-                                    <a href="bd_delete/facturacion-eliminar.php?id=<?=$res['fact_id'];?>&cte=<?=$_GET['cte'];?>" onClick="if(!confirm('Desea eliminar el registro?')){return false;}" data-toggle="tooltip" title="Eliminar"><i class="icon-remove-sign"></i></a>
+									<?php if( Modulos::validarRol(['261'], $conexionBdPrincipal, $conexionBdAdmin, $datosUsuarioActual, $configuracion) ) {?>
+                                		<a href="facturacion-editar.php?id=<?=$res['fact_id'];?>&cte=<?=$_GET['cte'];?>" data-toggle="tooltip" title="Editar"><i class="icon-edit"></i></a>&nbsp;
+									<?php }?>
+									<?php if( Modulos::validarRol(['58'], $conexionBdPrincipal, $conexionBdAdmin, $datosUsuarioActual, $configuracion) ) {?>
+                                    	<a href="bd_delete/facturacion-eliminar.php?id=<?=$res['fact_id'];?>&cte=<?=$_GET['cte'];?>" onClick="if(!confirm('Desea eliminar el registro?')){return false;}" data-toggle="tooltip" title="Eliminar"><i class="icon-remove-sign"></i></a>
+									<?php }?>
                                 </h4>
                                 </td>
 
                                 
                                 <td><?php 
 								while($pds = mysqli_fetch_array($Cpd, MYSQLI_BOTH)){
-									echo '<a href="productos-materiales.php?pdto='.$pds['prod_id'].'" title="Ver materiales">'.$pds['prod_nombre']."</a><br>";
+									if( Modulos::validarRol(['68'], $conexionBdPrincipal, $conexionBdAdmin, $datosUsuarioActual, $configuracion) ) {
+										echo '<a href="productos-materiales.php?pdto='.$pds['prod_id'].'" title="Ver materiales">'.$pds['prod_nombre']."</a><br>";
+									} else {
+										echo $pds['prod_nombre']."<br>";
+									}
 								}
 								?></td>
